On branch itemStyles # Changes to be committed: #	modified:   client/src/App/App.tsx #	modified:   client/src/App/api.ts #	deleted:    client/src/features/arts/PostItem.tsx #	modified:   client/src/features/categories/types.ts #	modified:   client/src/features/main/MainPage.tsx #	renamed:    client/src/features/arts/FormAddArt.tsx -> client/src/features/posts/FormAddArt.tsx #	renamed:    client/src/features/arts/PostPage.tsx -> client/src/features/posts/OnePostPage.tsx #	new file:   client/src/features/posts/PostItem.tsx #	renamed:    client/src/features/arts/PostsList.tsx -> client/src/features/posts/PostsListPage.tsx #	renamed:    client/src/features/arts/postsSlice.ts -> client/src/features/posts/postsSlice.ts #	new file:   client/src/features/posts/styles.css #	renamed:    client/src/features/arts/types.ts -> client/src/features/posts/types.ts #	modified:   client/src/redux/reducers/types.ts #	modified:   client/src/redux/store.ts
