+++ client/src/features/posts/styles.css
@@ -1,0 +1,29 @@
+.post-container {
+     max-width: 200px;
+    max-height: 200px;
+    /background-color: black;
+    flex: 1 1 auto;
+  margin: 5px 5px;
+  border: 3px solid red; 
+  position: relative; 
+    width: 50%;
+    overflow: hidden;
+    background: #2e5fa0;
+}
+.post-container::beforre {
+    content: "";
+    display: block;
+    padding-top: 100%;
+}
+.all-post-container {
+    display: flex;
+    flex-wrap: wrap;
+    margin: 0 -5px;
+    
+}
+.post-content {
+    color: white;
+}
+h1 {
+    color: white;
+}
