+++ client/src/redux/reducers/types.ts
@@ -1,13 +1,11 @@
-
-
-import type { Post, PostId } from '../../features/arts/types';
-import type{ Category } from '../../features/categories/types';
+import type { Post, PostId } from '../../features/posts/types';
+import type { Category } from '../../features/categories/types';
 import type { User } from '../../features/users/types';
 
 export type PostsState = {
   posts: Post[];
   error: string | undefined;
-//   loading: boolean;
+  //   loading: boolean;
 };
 
 export type AuthState = {
@@ -15,12 +13,10 @@ export type AuthState = {
 };
 
 export type Action =
-
   | { type: 'posts/load'; payload: Post[] }
   | { type: 'posts/add'; payload: Post }
   | { type: 'posts/remove'; payload: PostId }
-   | { type: 'categories/load'; payload: Category[] }
+  | { type: 'categories/load'; payload: Category[] }
   | { type: 'auth/sign-up'; payload: User }
   | { type: 'auth/sign-in'; payload: User }
   | { type: 'auth/logout' };
-
